fixing validation groups, move tag interface to top;
