+++ core-base/core-base/src/main/java/org/nanotek/beans/entity/SortNameBase.java
@@ -9,9 +9,11 @@ import javax.persistence.Inheritance;
 import javax.persistence.InheritanceType;
 import javax.persistence.Table;
 import javax.validation.constraints.NotBlank;
+import javax.validation.groups.Default;
 
 import org.nanotek.entities.BaseSortNameEntity;
 import org.nanotek.entities.MutableSortNameEntity;
+import org.nanotek.opencsv.CsvValidationGroup;
 
 @Entity
 @Table(name="sort_name_base",
@@ -31,7 +33,7 @@ MutableSortNameEntity<String>{
 
 	private static final long serialVersionUID = -950822256693332353L;
 
-	@NotBlank
+	@NotBlank(groups = {CsvValidationGroup.class,Default.class})
 	@Column(name="sort_name" , nullable=false, columnDefinition = "VARCHAR NOT NULL")
 	public String sortName;
