+++ core-base/core-base/src/main/java/org/nanotek/opencsv/CsvBaseProcessor.java
@@ -12,11 +12,11 @@ import org.nanotek.Base;
 import org.nanotek.BaseBean;
 import org.nanotek.IdBase;
 import org.nanotek.ImmutableBase;
+import org.nanotek.ProcessorBase;
 import org.nanotek.ValueBase;
 import org.nanotek.collections.BaseMap;
 import org.nanotek.opencsv.file.CsvFileItemConcreteStrategy;
 import org.nanotek.opencsv.task.CsvProcessorCallBack;
-import org.nanotek.processor.ProcessorBase;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.BeansException;
